global: search & detail layout
