+++ src/App.scss
@@ -1,7 +1,7 @@
 .__App__ {
   min-height: 100vh;
   .content {
-    padding: 0 50px;
+    padding: 0 40px;
     margin-top: 64px;
   }
 }
