+++ src/common/containers/AggregationFiltersContainer/AggregationFiltersContainer.scss
@@ -1,0 +1,4 @@
+.__AggregationFiltersContainer__ {
+  background: #fff;
+  padding: 10px;
+}
