+++ src/common/containers/AggregationFiltersContainer/index.js
@@ -1,0 +1,3 @@
+import AggregationFiltersContainer  from './AggregationFiltersContainer';
+
+export default AggregationFiltersContainer;
