+++ src/common/layouts/SearchLayout/SearchLayout.scss
@@ -1,0 +1,6 @@
+.__SearchLayout__ {
+  margin: 20px 0;
+  .alignRight {
+    text-align: right;
+  }
+}
