+++ src/common/layouts/SearchLayout/index.js
@@ -1,0 +1,3 @@
+import SearchLayout from './SearchLayout';
+
+export default SearchLayout;
